move dashboard name to synchronize name
