+++ src/EcclesiaCRM/Synchronize/DashboardItemInterface.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace EcclesiaCRM\Dashboard;
+namespace EcclesiaCRM\Synchronize;
 
 interface DashboardItemInterface {
-  
+
   //must be all one word - no spaces
   public static function getDashboardItemName();
 
   //when provided with the page name of the user context, return true or false if this item should be loaded / provided in AJAX updates.
   public static function shouldInclude($PageName);
-  
+
   // return a PHP array with all of the values to be passed to the renderer
   public static function getDashboardItemValue();
 }
